extract useMessages
