+++ app/src/Websocket.tsx
@@ -1,42 +1,20 @@
-import { useState, useEffect } from 'react';
 import { useParams, useLocation } from 'react-router-dom';
 import usePromise from 'react-promise-suspense';
 import { useAuth0 } from '@auth0/auth0-react';
-import useWebSocket, { ReadyState } from 'react-use-websocket';
+import { ReadyState } from 'react-use-websocket';
 import * as _ from 'lodash-es';
 
 import type { ITorrent } from './select/OptionsComponent';
-import { getMarker, getPrefix, getToken } from './utils';
+import { getMarker, getToken } from './utils';
 import type { components } from './schema';
 import { DisplayTorrent, RouteTitle } from './components';
+import { useMessages } from './components/websocket';
 
-type BaseRequest = components['schemas']['BaseRequest'];
 type StreamArgs = components['schemas']['StreamArgs'];
 
 function get(query: URLSearchParams, key: string): number | undefined {
   const value = query.get(key);
   return value ? parseInt(value, 10) : undefined;
-}
-
-function useMessages<T>(initMessage: BaseRequest) {
-  const base = getPrefix();
-  const url = `${base}/ws`;
-
-  const { sendJsonMessage, lastJsonMessage, readyState } = useWebSocket(url);
-
-  useEffect(() => {
-    sendJsonMessage(initMessage);
-  }, [sendJsonMessage, initMessage]);
-
-  const [messages, setMessages] = useState<T[]>([]);
-
-  useEffect(() => {
-    if (lastJsonMessage) {
-      setMessages((messages) => messages.concat([lastJsonMessage as T]));
-    }
-  }, [lastJsonMessage]);
-
-  return { messages, readyState };
 }
 
 function Websocket() {
